+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 0.3.9-dev+4.11.2014
 
 - Added the `scrollToLine` and `scrollToRow` methods to `Editor`.
-- Added the `getTokenAt` method to `EditSession` and the `Token` class.
+- Added the `getTokenAt` and `getTokens` methods to `EditSession` and the 
+`Token` class.
 
 ## 0.3.8+4.11.2014